requirements-offline.txt
# Notes:
#   The following dependencies are available on helios compute nodes

flake8
pytest
pytest-cov

tensorflow_gpu==2.0.0 # Keep fixed
pandas
jsonschema
tqdm
h5netcdf
